Include "explicit" tag in Deemix config file
I occasionally need to make sure I've got the "clean" version of a song (or I want the "explicit" one for personal listening).

This will show as "ITUNESADVISORY = 1" in the extended tags for a given file (0 for "clean" tracks). Including this tag will allow users to filter/tag explicit tracks using beets, making them easier to sort in a media player or DJ software.